-now only deletes old ParkingData if it has been sorted

+++ parkguidance-admin/src/main/java/com/neo/parkguidance/web/admin/pages/datalist/ParkingDataListFacade.java
@@ -86,8 +86,10 @@ public class ParkingDataListFacade {
 
         if (list != null) {
             for(ParkingData parkingData: list) {
-                deleted++;
-                parkingDataService.remove(parkingData);
+                if (Boolean.TRUE.equals(parkingData.getSorted())) {
+                    deleted++;
+                    parkingDataService.remove(parkingData);
+                }
             }
         }
         return deleted;
